Refactor loss function to calculate Mean Squared Error (MSE)
The `loss` function in `loss.hpp` has been refactored to calculate the Mean Squared Error (MSE) between two arrays. This measure is commonly used in regression analysis to assess the accuracy of a model. The function now takes pointers to the arrays of true values and predicted values, as well as the number of elements in the arrays. It calculates the sum of squared differences between the elements and returns the mean squared error.